[type: refactor] refactor envcd context build.
